requirements.txt
accelerate
diffusers==0.21.4
einops==0.6.1
ema-pytorch==0.2.3
numpy
opencv-python==4.8.1.78
pandas==2.1.2
Pillow
tqdm
transformers==4.34.0